chore: remove dead code — instinctllm pkg, Ollama backend, token bucket
Whole-program reachability analysis (golang.org/x/tools/cmd/deadcode) plus
go-list reverse-dependency checks identified these as unreachable from every
binary entrypoint. staticcheck ./... is clean, confirming no package-level
unused code remains after removal. Build + vet + full test suite (51 pkgs)
all green.

Removed:
- internal/instinctllm/ (entire package, 487 src + 564 test LOC): not in any
  binary's dependency tree; superseded by internal/llmclient
- internal/embed/ollama.go (OllamaClient, 12/12 funcs dead): Ollama embed
  backend abandoned for the Olla router + LiteLLMClient. The embed.Client
  interface it housed is extracted to internal/embed/client.go (the interface
  never belonged in a concrete impl file). Also drops the now-stale
  "Only Ollama is supported" package doc.
- internal/embed/token_bucket.go: NewEmbedTokenBucket/TryConsume/refill, all
  unreachable
- search/{decay,score}.go: ResetDecay{Interval,Config}ForTesting — zero refs
  even in _test.go

Net -1580 lines. Reversible via git.
