+++ internal/search/decay.go
@@ -22,11 +22,6 @@ var (
 	resolvedDecayInterval time.Duration
 )
 
-// ResetDecayIntervalForTesting resets the decay interval sync.Once so tests can inject different env var values.
-func ResetDecayIntervalForTesting() {
-	decayIntervalOnce = sync.Once{}
-}
-
 // resolveDecayInterval reads ENGRAM_DECAY_INTERVAL_HOURS and returns the
 // corresponding duration. If the env var is absent, invalid, or non-positive,
 // it returns defaultDecayInterval (8h) and logs a warning for invalid values.
